[front] feat: Add pptx typography feedbackloop
